update: portabletext/react library
